vibevoice-web/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
transformers>=4.51.3,<5.0.0
torch>=2.2.0
torchaudio>=2.2.0
accelerate>=0.29.0
huggingface_hub>=0.22.0
numpy>=1.26.0
scipy>=1.13.0
soundfile>=0.12.1
aiofiles>=23.2.1
sqlalchemy>=2.0.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
diffusers
git+https://github.com/microsoft/VibeVoice.git
